Applied review feedback
- Split out nested builders into separate module
- Improved documentation
- Added `LoadContext::load` thunk
